+++ app/services/quotes/collection/fetch_service.rb
@@ -6,15 +6,18 @@ module Quotes
       prepend BasicService
 
       def initialize(
-        from_moex_service: Fetching::MoexService
+        from_moex_service:    Fetching::MoexService,
+        from_tinkoff_service: Fetching::TinkoffService
       )
-        @from_moex_service = from_moex_service
+        @from_moex_service    = from_moex_service
+        @from_tinkoff_service = from_tinkoff_service
       end
 
       def call(source:, date:)
         @result =
           case source
           when Sourceable::MOEX then @from_moex_service.call(date: date).result
+          when Sourceable::TINKOFF then @from_tinkoff_service.call.result
           end
       end
     end