+++ CHANGELOG.md
@@ -19,3 +19,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - creating portfolios on account page
 - creating positions for specific portfolios
 - Portfolios::Cashes::Operation model
+- sync service for Tinkoff api